% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{add_worksheet_data}
\alias{add_worksheet_data}
\title{Writes a data frame to an Excel workbook with formatting}
\usage{
add_worksheet_data(
  wb,
  sheet_name,
  data,
  start_row,
  discrepancy_col = "Discrepancy Change",
  match_text = "Data match",
  finding_col = "Finding"
)
}
\arguments{
\item{wb}{an `openxlsx` `Workbook` object}

\item{sheet_name}{string, the sheet name to create}

\item{data}{data frame of the data for `sheet_name`}

\item{start_row}{integer, the row number at which to write the column header
of `data`}

\item{discrepancy_col}{a string, the column name for discrepancy change
tracking between report versions, default is `"Discrepancy Change"`. If set
to `NULL`, conditional formatting for discrepancy change will not be applied.}

\item{match_text}{a string, text that identifies a data match in
`discrepancy_col`, default is `"Data match"`}

\item{finding_col}{a string, the column name for reconciliation findings,
default is `"Finding"`. If set to `NULL`, conditional formatting for
discrepancy change will not be applied.}
}
\value{
a modified [openxlsx::createWorkbook()] object
}
\description{
Writes a data frame to the specified sheet in an Excel workbook and applies
formatting, conditional formatting, freeze panes, and filters.
}
\seealso{
[add_worksheet_header()], [setup_workbook()]
}
