% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{add_worksheet_header}
\alias{add_worksheet_header}
\title{Writes header data to Excel workbook}
\usage{
add_worksheet_header(wb, sheet_name, header, data)
}
\arguments{
\item{wb}{an `openxlsx` `Workbook` object}

\item{sheet_name}{string, the sheet name to create}

\item{header}{a list of strings where each element corresponds to one row in
the header}

\item{data}{data frame of the data for `sheet_name`}
}
\value{
an `openxlsx` `Workbook` object
}
\description{
Writes the meta data to the top rows of an Excel workbook and formats them
}
\seealso{
[add_worksheet_data()], [setup_workbook()]
}
