% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_code.R
\name{archive_code}
\alias{archive_code}
\title{Archive an R script}
\usage{
archive_code(code_fname, code_dir, code_archive_dir)
}
\arguments{
\item{code_fname}{a string, the file name of the code to archive}

\item{code_dir}{a string, the directory the code lives in}

\item{code_archive_dir}{a string, the directory the code should be copied to}
}
\value{
nothing
}
\description{
Copies the an R script to a data archive folder
}
