% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows_non_empty.R
\name{bind_rows_non_empty}
\alias{bind_rows_non_empty}
\title{Bind multiple data frames by rows ignore empty}
\usage{
bind_rows_non_empty(...)
}
\arguments{
\item{...}{data frames to combine}
}
\value{
a data frame
}
\description{
This is a improved version of [dplyr::bind_rows()] function.
To handle case that empty csv data file has different data structure,
this [bind_row_non_empty()] function omit data frame which are empty.
}
