% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_listing_to_previous.R
\name{categorize_finding_change}
\alias{categorize_finding_change}
\title{Compare a data frame listing to a previous version}
\usage{
categorize_finding_change(
  listing,
  prev_report,
  unique_row_id = "Unique_Row_ID",
  finding_col = "Finding",
  change_col = "Discrepancy Change"
)
}
\arguments{
\item{listing}{a data frame of with the new data}

\item{prev_report}{a data frame with the old data}

\item{unique_row_id}{a string, the name of the column that links unique row
IDs between `listing` and `prev_report`, usually a concatenation of other
column values that collectively make a unique ID; default is `"Unique_Row_ID"`}

\item{finding_col}{a string, the name of the column that describes row
discrepancies or issues, default is `"Finding"`}

\item{change_col}{a string, the name of the column to record changes; default
is `"Discrepancy Change"`}
}
\value{
a data frame
}
\description{
Compares a data frame listing to a previous version by annotating which rows
were deleted, changed, or are new.
}
\details{
An error will be raised if a row cannot be categorized into deleted, changed,
or new.
}
