% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_env.R
\name{cleanup_env}
\alias{cleanup_env}
\title{Clean-up global environment and search path}
\usage{
cleanup_env()
}
\value{
nothing
}
\description{
Clears all objects from the global environment and restarts R
}
\details{
Should be used after each report is run so R/R Studio is ready for the next
report
}
