% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_listing_to_previous.R
\name{compare_listing_to_previous}
\alias{compare_listing_to_previous}
\title{Compare a data frame listing to a previous version}
\usage{
compare_listing_to_previous(listing, prev_report)
}
\arguments{
\item{listing}{a data frame of with the new data}

\item{prev_report}{a data frame with the old data}
}
\value{
a data frame
}
\description{
Compares a data frame listing to a previous version by annotating which rows
were deleted, changed, or are new.
}
\details{
An error will be raised if a row cannot be categorized into deleted, changed,
or new.
}
