% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{create_excel_listing}
\alias{create_excel_listing}
\title{Make an Excel workbook}
\usage{
create_excel_listing(sheet_data, sheet_headers)
}
\arguments{
\item{sheet_data}{a named list of data frames containing the data for each
worksheet where the list names are the sheet names}

\item{sheet_headers}{a named list of lists where each name is a sheet
name in `wb` and each element is the header data for a worksheet in `wb`. If
`length(sheet_headers) == 1`, `sheet_headers` will be recycled for every
sheet in `wb`. `names(sheet_data)` must match `names(sheet_headers)`.}
}
\value{
an `openslxs` `Workbook` object
}
\description{
Creates an [openxlsx::createWorkbook()] object with one or more sheets and
populates header information and data for each sheet.
}
\seealso{
[setup_workbook()], [setup_multi_sheet_workbook()],
[populate_worksheet()], [populate_workbook()]
}
