% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Find ID duplicates}
\usage{
find_duplicates(df, keys, message, col = Discrepancy)
}
\arguments{
\item{df}{Data Frames to find duplicates}

\item{keys}{keys to determine duplicates}

\item{message}{content of column indicates duplication}

\item{col}{column name of generated column for duplication}
}
\value{
a modified version of `df`
}
\description{
Determines which observations of a data frame are duplicates based on the key
provided in the arguments. The function does not change the dimension of
the given data frame but instead creates a new column named by argument col.
The content of newly created column is specified by the message argument.
}
