% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_date_DDMMMYYYY}
\alias{format_date_DDMMMYYYY}
\title{Covert dates in YYYYMMDD format to DDMMMYYYY}
\usage{
format_date_DDMMMYYYY(my_date)
}
\arguments{
\item{my_date}{a string or Date in YYYYMMDD format}
}
\value{
a string in DDMMMYYYY format
}
\description{
Covert dates in YYYYMMDD format to DDMMMYYYY, not month will be in all caps.
}
