% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_date.R
\name{get_file_date}
\alias{get_file_date}
\title{Get file Generated Date}
\usage{
get_file_date(path)
}
\arguments{
\item{path}{absolute file path}
}
\value{
a Date
}
\description{
Return the date of a specific file generated date
}
