% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%neqna\%}
\alias{\%neqna\%}
\title{Operator for not equal comparison with NA handling}
\usage{
e1 \%neqna\% e2
}
\arguments{
\item{e1, e2}{vectors to compare; must be of equal length and have the same
class}
}
\value{
a vector of length equal to `e1` and `e2` containing logical values
}
\description{
Provides element-wise comparison of two vectors to determine if the values
are NOT equal. Comparison of two non-`NA` values is equivalent to the result
of `value1 == value2`.`NA` compared to a non-`NA` values will return `TRUE`
and `NA` compared to `NA` will return `FALSE`.
}
\examples{
# the below will return `c(FALSE, TRUE, TRUE, FALSE)`
c(1, 1, 1, NA) \%neqna\% c(1, 2, NA, NA)

}
