% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data}
\usage{
import_data(source_dir, filetype, filenames = c(), ...)
}
\arguments{
\item{source_dir}{a string of directory contains dataset files}

\item{filetype}{a string of file type, support sas7bdat, csv, rds, xlsx}

\item{filenames}{a vector contains file names}

\item{...}{additional arguments passed to one of the reading funtions}
}
\value{
a list of data frames
}
\description{
Import data sets within a folder based on the file type and file name
}
