% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listingutils-package.R
\docType{package}
\name{listingutils-package}
\alias{listingutils}
\alias{listingutils-package}
\title{listingutils: Utilities for Data Listings}
\description{
Provides utility functions for statistical programmers to create data listings in MS Excel.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/skgithub14/listingutils}
  \item Report bugs at \url{https://github.com/skgithub14/listingutils/issues}
}

}
\author{
\strong{Maintainer}: Stephen Knapp \email{stephen@knappconsultingllc.com} (\href{https://orcid.org/0000-0002-5101-4555}{ORCID})

Authors:
\itemize{
  \item Jason Zhang \email{jason.zhang@edetek.com}
}

Other contributors:
\itemize{
  \item Inga Lin \email{inga.lin@oncology.sumitomo-pharma.com} [contributor]
  \item Sylvia Chen \email{sylvia.chen@oncology.sumitomo-pharma.com} [contributor]
}

}
\keyword{internal}
