% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_record_by_joining.R
\name{match_record_by_joining}
\alias{match_record_by_joining}
\title{This function is designed to support DM data reconciliation report.
To reconciliate the date, EDC records and vendor records need to be matched
together. Most of the time, there is no direct id join key and instead
requires composite join key such as subject, visit, date, time, timepoint.
In the case, some of the composite key is not perfectly matches, this
function try different set of composite keys to match the records.}
\usage{
match_record_by_joining(
  df1,
  df2,
  bys,
  distinct_keys,
  suffix = c(".edc", ".vendor"),
  na_matches = "never"
)
}
\arguments{
\item{df1}{Data Frames to match (usually EDC data)}

\item{df2}{Data Frame to be mathced (usually vendor data)}

\item{bys}{Sets of composite keyes used to match}

\item{distinct_keys}{a character vector of distinct keys}

\item{suffix}{a character vector of length two of column suffixes to append
to `df1` and `df2` columns, respectively}

\item{na_matches}{whether to match `NA` values, see the dplyr join functions}
}
\value{
a data frame
}
\description{
This function is designed to support DM data reconciliation report.
To reconciliate the date, EDC records and vendor records need to be matched
together. Most of the time, there is no direct id join key and instead
requires composite join key such as subject, visit, date, time, timepoint.
In the case, some of the composite key is not perfectly matches, this
function try different set of composite keys to match the records.
}
