% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{populate_workbook}
\alias{populate_workbook}
\title{Populate workbook data for all worksheets}
\usage{
populate_workbook(wb, sheet_names, sheet_data, sheet_headers)
}
\arguments{
\item{wb}{an `openxlsx` `Workbook` object}

\item{sheet_names}{a list of worksheet names}

\item{sheet_data}{a named list of data frames containing the data for each
worksheet; the names of `sheet_data` should match the contents of
`sheet_names` and `sheet_data` should have the same length as `sheet_names`}

\item{sheet_headers}{a named list of lists where each name is a sheet
name in `wb` and each element is the header data for a worksheet in `wb`. The
names of `sheet_headers` must match the elements of `sheet_names` and
`length(sheet_headers)` must be equal to `length(sheet_name)` or `1`. If
`length(sheet_headers) == 1`, `sheet_headers` will be recycled for every
sheet in `wb`.}
}
\value{
an `openxlsx` Workbook object
}
\description{
Wraps [populate_worksheet()] to populate all sheets in an `openxlsx` Workbook
object.
}
\seealso{
[populate_worksheet()], [setup_multi_sheet_workbook()]
}
