% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{populate_worksheet}
\alias{populate_worksheet}
\title{Populates a worksheet in an Excel workbook with header info and data}
\usage{
populate_worksheet(wb, sheet_name, data, header)
}
\arguments{
\item{wb}{an `openxlsx` `Workbook` object}

\item{sheet_name}{string, the sheet name to create}

\item{data}{data frame of the data for `sheet_name`}

\item{header}{a list of strings where each element corresponds to one row in
the header}
}
\value{
an `openxlsx` Workbook object
}
\description{
Wraps [add_worksheet_header()] and [add_worksheet_data()] to populate an
entire sheet's worth of content.
}
\details{
The sheet must already have been created in the workbook
}
\seealso{
[add_worksheet_header()], [add_worksheet_data()], [setup_workbook()]
}
