% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{round2}
\alias{round2}
\title{Apply normal rounding rules}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{a numeric value to round}

\item{digits}{the number of digits to round to, integer}
}
\value{
a rounded numeric value
}
\description{
This is a function to round numeric vectors to the chosen number of
decimal places, following the rule "Round to nearest, ties away from
zero
}
\details{
The function was adapted from
\url{https://stackoverflow.com/questions/12688717/round-up-from-5}
}
