% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{setup_workbook}
\alias{setup_workbook}
\title{Set-up Excel workbook with one worksheet}
\usage{
setup_workbook(sheet_name, data, col_widths = 19)
}
\arguments{
\item{sheet_name}{string, the sheet name to create}

\item{data}{data frame of the data for `sheet_name`}

\item{col_widths}{a numeric value or `"auto"`, the column widths for all
worksheets; default is `19`.}
}
\value{
an `openxlsx` `Workbook` object
}
\description{
Creates an [openxlsx::createWorkbook()] object, adds a worksheet, and sets
column widths to 19. This function should be used when creating workbooks
with one and only one sheet. For multi-sheet workbooks, see
[setup_multi_sheet_workbook()].
}
\details{
No data is written to the workbook. The `data` argument is just for
configuring options for the correct number of columns.
}
\seealso{
[setup_multi_sheet_workbook()]
}
