% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{add_worksheet_data}
\alias{add_worksheet_data}
\title{Writes a data frame to Excel workbook with formatting}
\usage{
add_worksheet_data(
  wb,
  sheet_name,
  dat,
  start_row,
  disrepancy_col = "Discrepancy Change",
  match_text = "Data match",
  finding_col = "Finding"
)
}
\arguments{
\item{wb}{an [openxlsx::createWorkbook()] object}

\item{sheet_name}{string, the sheet name to write to}

\item{dat}{data frame, the data to write}

\item{start_row}{integer, the row number at which the write the column header
of `dat`}

\item{disrepancy_col}{a string, the column name for discrepancy change
tracking between report versions, default is `"Discrepancy Change"`}

\item{match_text}{a string, text that identifies data match in
`disrepancy_col`, default is `"Data match"`}

\item{finding_col}{a string, the column name for reconciliation findings,
default is `"Finding"`}
}
\value{
a modified [openxlsx::createWorkbook()] object
}
\description{
Writes a data frame to the specified sheet in an Excel workbook and applies
formatting, conditional formatting, freeze panes, and filters.
}
