% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{add_worksheet_header}
\alias{add_worksheet_header}
\title{Writes header data to Excel workbook}
\usage{
add_worksheet_header(wb, sheet_name, h_data, dat)
}
\arguments{
\item{wb}{an `openxlsx::createWorkbook()` object}

\item{sheet_name}{a string, the sheet name in `wb` to write to}

\item{h_data}{a list where each element corresponds to one row in the header}

\item{dat}{a data frame of the data that will be written below the header}
}
\value{
a modified `openxlsx::createWorkbook()` object
}
\description{
Writes the meta data to the top rows of an Excel workbook and formats them
}
