% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{setup_multi_sheet_workbook}
\alias{setup_multi_sheet_workbook}
\title{Set-up Excel workbook with multiple worksheets}
\usage{
setup_multi_sheet_workbook(sheet_names, sheet_data)
}
\arguments{
\item{sheet_names}{a list of worksheet names}

\item{sheet_data}{a named list of data frames containing the data for each
worksheet; the names of `sheet_data` should match the contents of
`sheet_names` and `sheet_data` should have the same length as `sheet_names`}
}
\value{
an `openxlsx` `Workbook` object
}
\description{
Creates an [openxlsx::createWorkbook()] object and adds worksheets.
}
\details{
No data is written to the workbook. The `data` argument is just for
configuring options for the correct number of columns. The column widths are
also set to 19.
}
\seealso{
[setup_workbook()]
}
