% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{setup_workbook}
\alias{setup_workbook}
\title{Set-up Excel workbook with standard options}
\usage{
setup_workbook(sheet_name, dat)
}
\arguments{
\item{sheet_name}{string, the sheet name to create for `dat`}

\item{dat}{data frame of the data for the sheet}
}
\value{
a `openxlsx::createWorkbook()` workbook object
}
\description{
Creates a `openxlsx::createWorkbook()` object, adds a worksheet, sets column
widths to auto and configures options for borders, dates and maximum column
widths.
}
\details{
No data is written to the workbook. The `dat` argument is just for configuring
options for the correct number of columns.
}
